requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sentence-transformers
faiss-cpu
openai
transformers
torch
pdfplumber
python-magic
cross-encoder